Add queries.sql with first query, first attempt
This query attempts to select "the most popular three articles of all
time" from the database. This query selects the titles of the articles
and the 'total' rows for each article title. The query selects this data
from the tables 'articles' and 'log'. These two tables are joined by
comparing log.path to the concatenation of '/article/' and
articles.slug, since the only difference between the values of these
two columns is the string, '/article/'. This concatenation also excludes
any invalid path. The query groups rows by article title and orders rows
by 'total' in descending order. The query limits the results to 3 rows.