Add third query, version 1
This query is incomplete. It is the first draft of a query that will
answer the question, "On which days did more than 1% of requests lead to
errors?" The query joins two subqueries. The first subquery selects the
date (truncating the timestamp to the day) and the count of all requests
errors per each date. The second subquery selects the date (also
truncating the timestamp) and the count of all requests. Both subqueries
select from the 'log' table. The query joins the subqueries by date. The
query selects the date, error count, and total count from the joining of
the subqueries. So far, the query does not accomplish more than this. In
the future, the error count may be divided by the total count to return
the percentage of errors per day.