Introduce options for saving answers in text file
All options are commented out because they don't work. They won't
create new lines, ruining the text-based tables. This may be because
Python 3.5 is running the code, and I know it works with other projects
on 3.7, but I am unable to test this theory. I'll do some research and
possibly come back to this part of the project. It may not even be
necessary to complete.